Serialize publish messages directly to buffer
Here I moved the serialization to the left side of the queue, avoiding having to rent memory and copy that memory to a buffer at the other end.
The only case where memory is rented now is if the message is larger than the predefined buffers. The overhead of renting memory to serialize smaller messages was a bottleneck.
Memory allocation on the send path is well controlled now, since even renting from the pool, at high volume scenarios, would end up allocating since the pool might be exhausted.
Another factor is that now different publishes only contend to acquire a write slot, but can serialize in parallel to the buffer, and even after the buffer enters the queue. The socket sender task will commit and seal the buffer at the last possible moment.

This increased throughput by ~2x at least.